camera: allow loading files with text relocations from /system
the qcom camera driver needs to load a shared library from /system
which contains a text relocation. Allow it.

Due to bug 20013628, SELinux policy was inappropriately treating
an execmod denial as an execmem denial. Move to using a proper
execmod denial and get rid of execmem.

Addresses the following denial:
  avc: denied { execmod } for pid=387 comm="mm-qcamera-daem" path="/system/vendor/lib/libmmcamera_faceproc.so" dev="mmcblk0p41" ino=1871 scontext=u:r:camera:s0 tcontext=u:object_r:system_file:s0 tclass=file permissive=0

Bug: 20013628
Change-Id: I6611df77b9cbf2a78ed595d19fc80d63aa837183